Add authorId validation to DeleteQuestionUseCase

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -1,6 +1,7 @@
 import { QuestionsRepository } from '../repositories/questions-repository'
 
 interface DeleteQuestionUseCaseRequest {
+  authorId: string
   questionId: string
 }
 
@@ -11,11 +12,16 @@ export class DeleteQuestionUseCase {
 
   async execute({
     questionId,
+    authorId,
   }: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
     const question = await this.questionsRepository.findById(questionId)
 
     if (!question) {
       throw new Error('Question not found')
+    }
+
+    if (question.authorId.toString() !== authorId) {
+      throw new Error('Not allowed')
     }
 
     await this.questionsRepository.delete(question)
